Move all models into a new subdirectory models/models.
